A.E : Mapper Modifications

+++ src/main/java/com/bagile/tms/entities/TmsTraffic.java
@@ -6,22 +6,26 @@ import java.util.Date;
 @Entity
 @Table(name="tms_roadTraffic",uniqueConstraints = @UniqueConstraint(columnNames = {"tms_roadTrafficId"}))
 public class TmsTraffic {
-    private int tmsTrafficId;
+    private long tmsTrafficId;
     private boolean tmsTrafficActive;
     private Date tmsTrafficDate;
+
+
 
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq")
     @SequenceGenerator(name = "seq")
     @Column(name = "tms_trafficId", unique = true, nullable = false, precision = 10, scale = 0)
-    public int getTmsTrafficId() {
+
+    public long getTmsTrafficId() {
         return tmsTrafficId;
     }
 
-    public void setTmsTrafficId(int tmsTrafficId) {
+    public void setTmsTrafficId(long tmsTrafficId) {
         this.tmsTrafficId = tmsTrafficId;
     }
+
 
     @Column(name = "tms_trafficactive")
     public boolean isTmsTrafficActive() {
